backend/infrastructure/elasticsearch/requirements.txt
elasticsearch>=8.13.0,<9
psycopg2-binary>=2.9.9
